Move FunctionalTestCase to testing.py
